Refactor, cleaning up code a bit before fixing / writing tests
