+++ td.vue/src/service/migration/labels.js
@@ -1,0 +1,11 @@
+const getText = (cell, label) => {
+    let text = label.attrs.text.text;
+    if (cell.protocol) {
+        text = `${text} (${cell.protocol})`;
+    }
+    return text;
+};
+
+export default {
+    getText
+};
